+++ config/config.go
@@ -152,8 +152,8 @@ type RecommendConfig struct {
 	ColIndexFitEpoch             int                `mapstructure:"collaborative_index_fit_epoch"`
 	EnableClickThroughPrediction bool               `mapstructure:"enable_click_through_prediction"`
 	EnableReplacement            bool               `mapstructure:"enable_replacement"`
-	PositiveReplacementDecay     float32            `mapstructure:"positive_replacement_decay"`
-	ReadReplacementDecay         float32            `mapstructure:"read_replacement_decay"`
+	PositiveReplacementDecay     float64            `mapstructure:"positive_replacement_decay"`
+	ReadReplacementDecay         float64            `mapstructure:"read_replacement_decay"`
 	exploreRecommendLock         sync.RWMutex
 }
